Refactor + listen to npc_spawns + reformat response of websocket
