+++ overlore/graphql/constants.py
@@ -2,9 +2,10 @@ from enum import Enum
 
 
 class EventType(Enum):
-    ORDER_ACCEPTED_EVENT = "0x20e86edfa14c93309aa6559742e993d42d48507f3bf654a12d77a54f10f8945"
-    TRANSFER_EVENT = "0x99cd8bde557814842a3121e8ddfd433a539b8c9f14bf31ebf108d12e6196e9"
+    ORDER_ACCEPTED = "0x20e86edfa14c93309aa6559742e993d42d48507f3bf654a12d77a54f10f8945"
+    TRANSFER = "0x99cd8bde557814842a3121e8ddfd433a539b8c9f14bf31ebf108d12e6196e9"
     COMBAT_OUTCOME = "0x1736c207163ad481e2a196c0fb6394f90c66c2e2b52e0c03d4a077ac6cea918"
+    NPC_SPAWNED = "0x24334a7bb6ddb29a34bf9e534498505917deb1101efad4807ae7b2dda50f6b6"
 
 
 class Queries(Enum):
@@ -74,12 +75,12 @@ class Queries(Enum):
             }
         }
         """
-    ORDER_ACCEPTED_EVENT_QUERY = EVENTS.format(event_hash=EventType.ORDER_ACCEPTED_EVENT.value)
+    ORDER_ACCEPTED_EVENT_QUERY = EVENTS.format(event_hash=EventType.ORDER_ACCEPTED.value)
     COMBAT_OUTCOME_EVENT_QUERY = EVENTS.format(event_hash=EventType.COMBAT_OUTCOME.value)
 
 
 class Subscriptions(Enum):
-    EVENT_EMITTED = """
+    KEY_BASED_SUB_TEMPLATE = """
           subscription {{
                 eventEmitted(keys: ["{event_hash}"]) {{
                     id
@@ -99,5 +100,6 @@ class Subscriptions(Enum):
                 }
             }
         """
-    ORDER_ACCEPTED_EVENT_EMITTED = EVENT_EMITTED.format(event_hash=EventType.ORDER_ACCEPTED_EVENT.value)
-    COMBAT_OUTCOME_EVENT_EMITTED = EVENT_EMITTED.format(event_hash=EventType.COMBAT_OUTCOME.value)
+    ORDER_ACCEPTED = KEY_BASED_SUB_TEMPLATE.format(event_hash=EventType.ORDER_ACCEPTED.value)
+    COMBAT_OUTCOME = KEY_BASED_SUB_TEMPLATE.format(event_hash=EventType.COMBAT_OUTCOME.value)
+    NPC_SPAWNED = KEY_BASED_SUB_TEMPLATE.format(event_hash=EventType.NPC_SPAWNED.value)
